Gem First Draft of Post and Reply Routes Complete

+++ Candid/app/api/reply_routes.py
@@ -1,21 +1,30 @@
 from flask import Blueprint, jsonify, request
-from flask_login import current_user
+from flask_login import current_user, login_required
 from sqlalchemy.orm import relationship, sessionmaker, joinedload, load_only
-from app.models import db, Reply
+from app.models import db, Reply, User
 
 reply_routes = Blueprint('reply', __name__)
 
 #Still Need to:
-    # Test Create, Edit, Delete routes and adjust code
-    # Include Authenticat and Authorization capability
-    # Provide Validation and Error handling
+    # Need to return Replies ordered by date
+    # Test Create, Edit, Delete routes and adjust code - 
+    # Include Authenticate/Authorization capability - 
+        # Need to make sure user is logged in in order to be complete requests
+        # Need to make sure query exists
+        # Need to make sure user owns post and authorized to make changes
+    # Provide Validation and Error handling - Will Complete on Front End
 
 # Get all replies under the comment
 @reply_routes.route('/comments/<int:id>/replies')
 def get_replies_for_comment(id):
 
-    replies = Reply.query.filter(Reply.comment_id == id).options(joinedload(Reply.user).options(load_only('id', 'username', 'preview_image'))).all()
-    
+    replies = Reply.query.filter(Reply.comment_id == id).options(joinedload(Reply.user).options(load_only('id', 'username', 'preview_image'))).order_by(Reply.created_at).all()
+    if not replies:
+        return {
+            "message": "Replies couldn't be found",
+            "statusCode": 404
+            }, 404 
+
     return {
         "Replies" : [
             {
@@ -37,6 +46,7 @@ def get_replies_for_comment(id):
 
 # Create a Reply under the comment
 @reply_routes.route('/comments/<int:id>/replies', methods=["POST"])
+@login_required
 def create_new_reply(id):
 
     currentuser = current_user.to_dict()
@@ -58,11 +68,26 @@ def create_new_reply(id):
 
 # Edit a Reply
 @reply_routes.route('/replies/<int:id>', methods=["PUT"])
+@login_required
 def edit_reply(id):
+
+    currentuser = current_user.to_dict()
+    user_id = currentuser['id']
     
     edit_reply = request.json['reply']
 
-    reply = Reply.query.filter(Reply.id == id).first()
+    reply = Reply.query.filter(Reply.id == id).one_or_none()
+    if not reply:
+        return {
+            "message": "Reply couldn't be found",
+            "statusCode": 404
+            }, 404
+
+    if not (reply.user_id == user_id):
+        return {
+            "message": "Forbidden",
+            "statusCode": 403
+            }, 403
     
     reply.reply = edit_reply
 
@@ -73,14 +98,29 @@ def edit_reply(id):
 
 # Delete a Reply
 @reply_routes.route('/replies/<int:id>', methods=["DELETE"])
+@login_required
 def delete_reply(id):
+
+    currentuser = current_user.to_dict()
+    user_id = currentuser['id']
     
-    reply = Reply.query.filter(Reply.id == id).first()
+    reply = Reply.query.filter(Reply.id == id).one_or_none()
+    if not reply:
+        return {
+            "message": "Reply couldn't be found",
+            "statusCode": 404
+            }, 404
+
+    if not (reply.user_id == user_id):
+        return {
+            "message": "Forbidden",
+            "statusCode": 403
+            }, 403
     
     db.session.delete(reply)
     db.session.commit()
 
     return {
         "message": "Successfully deleted",
-        "statusCode": 404
+        "statusCode": 200
     }
